Initial mini cleanup and made some notes on usage of BaseStationTimer… (#3253)
* Initial mini cleanup and made some notes on usage of BaseStationTimer class
- Remove GetRunTime call from ITIckTimeProvider (not used in Cozmo or OD)
- Remove unused GetTimeSinceLastTickInNanoSeconds
- Add comments (temporary!) on usage of calls in BaseStationTimer for both Cozmo and OD

* Change some system clock to steady clock

- The key change here is really in the main update where we used to call std::chrono::system_clock and now we call std::chron::steady_clock.  This avoid major ‘time elapsed’ differences if the user (or system, due to daylight saving time) changes the date/time on the device.
- Also a couple of other places where steady_clock makes more sense than system_clock.
- Also address some minor PR feedback

* Moar on time update
- More cleanup in basestation timer
- CozmoEngine::Update now takes a double, not a float, because previously we had a double, converted to a float, and then right back to a double again, thus losing precision.

* More timer usage fixes and optimizations
- Make some constants for robot off treads state TimeStamp_t instead of float, so we’re not casting at runtime
- Fix bugs in BehaviourReactToPyramid and BehaviorReactToStackOfCubes:  Call GetCurrentTimeInSeconds rather than GetCurrentTimeStamp, which returns milliseconds
- Added ‘const’ to some local vars found while sweeping usage of GetCurrentTimeStamp
- In GetCurrentTimeStamp, use the float and divide by 1M instead of the double and dividing by 1B
- Remove cruft from BaseStationTimer

* Use float for time in most cases
- BaseStationTimer now maintains time as a float (seconds), double (seconds), and BaseStationTime_t (unsigned long long int) (nanoseconds)
- Pass a BaseStationTime_t (nanoseconds) up to BaseStationTimer::UpdateTime from the cozmo loops
- Visit all ~130 callers of GetCurrentTimeInSeconds, since now that function returns a float, not a double:
  * Convert usage to use floats instead of doubles; this saves many runtime casts and simplifies the code
  * Remove a lot of casts no longer needed
  * Add a couple of casts now needed
  * Add const to several local vars that temporarily store the current time
- Remove a bit of unused code
- Add comments on two bugs in PanTiltTracker, verified as bugs by Andrew, but that code is unused

* More changes
- double -> float in a few more places, including the times stored in MoodManager
- A few cleanups based on feedback, like adding ‘_secs’ to some names to indicate the time unit
- Use numeric_cast instead of static_cast in the previous commit
- In sending duration of last AI goal, PRINT_NAMED_ERROR if calculated time was negative
- Remove temporary usage comments (notes) in BaseStationTimer class
- Fixes for PanTiltTracker

* Mini: Put steady_clock back into webots update
- This small change was lost in my merge (there’s probably a better way of merging than I was doing)

+++ basestation/include/anki/cozmo/basestation/actions/trackingActions.h
@@ -53,7 +53,7 @@ public:
   // Set how long the tracker will run without seeing whatever it is trying to
   // track. After this, it will complete "successfully".
   // Set to 0 to disable timeout (default).
-  void SetUpdateTimeout(double timeout_sec) { _updateTimeout_sec = timeout_sec; }
+  void SetUpdateTimeout(float timeout_sec) { _updateTimeout_sec = timeout_sec; }
 
   // Tells this action to keep running until another action (being run separately) stops. As soon as this
   // other action completes, this action will complete as well
@@ -113,8 +113,8 @@ protected:
 private:
   
   Mode     _mode = Mode::HeadAndBody;
-  double   _updateTimeout_sec = 0.;
-  double   _lastUpdateTime = 0.;
+  float    _updateTimeout_sec = 0.0f;
+  float    _lastUpdateTime = 0.0f;
   Radians  _panTolerance  = POINT_TURN_ANGLE_TOL;
   Radians  _tiltTolerance = HEAD_ANGLE_TOL;
   Radians  _maxHeadAngle  = MAX_HEAD_ANGLE;
